Removed Eager execution from TF2X. Updated TF2-models and some minor changes

+++ scripts/training/opening lead/keras/binary_lead_keras_suit_nt.py
@@ -71,7 +71,7 @@ class DealData(object):
                 # we only generate data for the opening lead hand
                 v_we = self.vuln_ns if hand_ix % 2 == 0 else self.vuln_ew
                 v_them = self.vuln_ew if hand_ix % 2 == 0 else self.vuln_ns
-                vuln = np.array([[v_we, v_them]], dtype=np.float32)
+                vuln = np.array([[v_we, v_them]], dtype=np.float16)
                 hcp = self.hcp[hand_ix]
                 shape = self.shapes[hand_ix]
