+++ src/bidding/bidding.py
@@ -55,17 +55,17 @@ def encode_bid(bid, alert_supported=False):
         alert = "*" in bid
         #Replace any alerts
         bid = bid.replace("*",'')
-        bid_one_hot = np.zeros((1, 41), dtype=np.float32)
+        bid_one_hot = np.zeros((1, 41), dtype=np.float16)
         bid_one_hot[0, 40] = 1 if alert else 0
         bid_one_hot[0, BID2ID[bid]] = 1
     else:
         bid = bid.replace("*",'')
-        bid_one_hot = np.zeros((1, len(BID2ID)), dtype=np.float32)
+        bid_one_hot = np.zeros((1, len(BID2ID)), dtype=np.float16)
         bid_one_hot[0, BID2ID[bid]] = 1
     return bid_one_hot
 
 def get_input(lho_bid, partner_bid, rho_bid, hand, v_we, v_them):
-    vuln = np.array([[v_we, v_them]], dtype=np.float32)
+    vuln = np.array([[v_we, v_them]], dtype=np.float16)
     return np.concatenate((vuln, encode_bid(lho_bid), encode_bid(partner_bid), encode_bid(rho_bid), hand), axis=1)
 
 def is_contract(bid):
